+++ src/imio/smartweb/common/upgrades.py
@@ -10,3 +10,8 @@ PROFILEID = "profile-imio.smartweb.common:default"
 def reload_registry(context):
     portal_setup = api.portal.get_tool("portal_setup")
     portal_setup.runImportStepFromProfile(PROFILEID, "plone.app.registry")
+
+
+def reload_actions(context):
+    portal_setup = api.portal.get_tool("portal_setup")
+    portal_setup.runImportStepFromProfile(PROFILEID, "actions")